simple_rag/requirements.txt
numpy
tiktoken
openai
python-dotenv